Attempting to run pretty 2.0.1 in [babashka](https://github.com/babashka/babashka) yields an error:
```
   (Namespace/find sym)
    ^--- No matching method find found taking 1 args
```

This has been introduced in a [recent change](https://github.com/clj-commons/pretty/pull/95).
Previous versions (I checked 2.0) are usable from babashka.

This is because `Namespace/find` is not compiled in babashka (probably considered an implementation detail).
But `clojure.core/find-ns` is included and is equivalent

https://github.com/clojure/clojure/blob/38bafca9e76cd6625d8dce5fb6d16b87845c8b9d/src/clj/clojure/core.clj#L4126-L4130

This change proposes to replace `Namespace/find` with `clojure.core/find-ns`.